+++ src/include/access/xlog_internal.h
@@ -46,8 +46,8 @@ typedef struct BkpBlock
 	RelFileNode node;			/* relation containing block */
 	ForkNumber	fork;			/* fork within the relation */
 	BlockNumber block;			/* block number */
-	uint16		hole_offset;	/* number of bytes before "hole" */
-	uint16		hole_length;	/* number of bytes in "hole" */
+	uint32		hole_offset;	/* number of bytes before "hole" */
+	uint32		hole_length;	/* number of bytes in "hole" */
 
 	/* ACTUAL BLOCK DATA FOLLOWS AT END OF STRUCT */
 } BkpBlock;
